organização-do-codigo
Adicionado as funções deletar e update no status de consulta. Organizei o código separando em funções gerais no 'mylib.c", medicos em 'medico.c", pacientes em "paciente.c" e consultas em "consulta.c"
