+++ mylib.c
@@ -4,225 +4,6 @@
 #include <stdlib.h>
 #include "mylib.h"
 
-void add_paciente(VetPacientes *pacientes){
-    FILE *file;
-    char nome[64];char contato[64]; char choice;
-    
-    int flag = realocar_pacientes(pacientes);
-    
-    if(flag == 0){
-        return;
-    }
-    
-    do{printf("Digite o nome do paciente: ");
-        getchar();
-        fgets(nome,64, stdin);
-        nome[strcspn(nome, "\n")] = '\0';
-    }while(verify_name(nome)==0);
-    
-    do{printf("Digite o e-mail do paciente: ");
-    fgets(contato,64, stdin);
-    contato[strcspn(contato, "\n")] = '\0';
-    }while(verify_email(contato)==0);
-    
-    	
-	int maior = 0;
-	if(pacientes->qtd > 0) {
-		for(int i = 0; i<pacientes->qtd; i++) {
-			if(pacientes->itens[i].id > maior) {
-				maior = pacientes->itens[i].id;
-			}
-		}
-	}
-
-
-	printf("Adicionar o paciente %s, contato: %s com o id: %d? (y/n)\n", nome, contato, maior+1);
-    while(1){
-        scanf(" %c", &choice);
-    
-    if(choice == 'n') return;
-    if(choice=='y')break;
-    }
-
-	printf("Adicionando...\n\n");
-
-
-	file = fopen("pacientes.txt","a"); //a-append
-
-	if(file == NULL) {
-		printf("Erro ao abrir arquivo");
-		return;
-	}
-
-	//fprintf é uma função para escrever no arquivo.
-	fprintf(file, "%s | %s | %d\n", nome, contato, maior+1);
-
-	fclose(file);
-
-	strcpy(pacientes->itens[pacientes->qtd].nome, nome);
-	strcpy(pacientes->itens[pacientes->qtd].contato, contato);
-	pacientes->itens[pacientes->qtd].id = maior+1;
-
-	pacientes->qtd++;
-
-	printf("Paciente cadastrado.\n");
-
-
-}
-
-int search_paciente(VetPacientes *pacientes) {
-	int id, i;
-	if(pacientes->qtd == 0) {
-		printf("Não existe pacientes cadastrados.\n");
-		return -2;
-	}
-	printf("Digite o id do paciente: ");
-	scanf("%d", &id);
-    for(i=0; i<pacientes->qtd; i++){
-        if(pacientes->itens[i].id == id){
-            printf("id: %d nome: %s contato: %s\n", pacientes->itens[i].id, pacientes->itens[i].nome, pacientes->itens[i].contato);
-            return i;
-        }
-    }
-    printf("Paciente não encontrado no sistema.\n");
-    return -1;
-}
-
-void read_pacientes(VetPacientes *pacientes) {
-	FILE *file;
-
-	int i = 0;
-	file = fopen("pacientes.txt", "r");
-
-	if(file == NULL) {
-		pacientes->qtd = 0;
-		pacientes->cap = 10;
-		pacientes->itens = (Paciente *) malloc(sizeof(Paciente) * pacientes->cap);
-
-		if(pacientes->itens == NULL) {
-			printf("Erro de memoria\n");
-			return;
-		}
-
-		file = fopen("pacientes.txt","w");
-
-		if(file == NULL) {
-			printf("Erro ao abrir o arquivo\n");
-		}
-		fclose(file);
-
-		return;
-	}
-
-	pacientes->qtd = contar_linhas(file);
-	pacientes->cap = pacientes->qtd + 10;
-	pacientes->itens = (Paciente *) malloc(sizeof(Paciente) * pacientes->cap);
-
-	if(pacientes->itens == NULL) {
-		printf("Erro ao realocar o vetor\n");
-		fclose(file);
-		return;
-	}
-
-
-	while(fscanf(file, "%63[^|] | %s | %d\n", &pacientes->itens[i].nome, &pacientes->itens[i].contato, &pacientes->itens[i].id) != EOF) {
-		i++;
-	}
-	pacientes->qtd = i;
-
-	fclose(file);
-}
-
-int realocar_pacientes(VetPacientes *pacientes) {
-	if(pacientes->qtd == pacientes->cap) {
-		pacientes->itens = (Paciente *) realloc(pacientes->itens, sizeof(Paciente) * (pacientes->cap+10));
-		if(pacientes->itens == NULL) {
-			printf("Erro de memoria!\n");
-			return 0;
-		}
-		pacientes->cap += 10;
-	}
-	return 1;
-}
-
-void remover_paciente(VetPacientes *pacientes){
-    int id = search_paciente(pacientes);
-    char choice;
-    if(id == -1){return;}
-    printf("Deseja remover esse paciente? (y/n)\n");
-    scanf(" %c", &choice);
-    if(choice == 'n')return;
-    
-    FILE *file;
-    
-    for(int i = id; i<pacientes->qtd-1; i++){
-        pacientes->itens[i] = pacientes->itens[i+1];
-    }
-    pacientes->qtd--;
-    
-    file = fopen("pacientes.txt", "w");
-    
-    for(int i = 0; i<pacientes->qtd; i++){
-        fprintf(file, "%s | %s | %d\n", pacientes->itens[i].nome, pacientes->itens[i].contato, pacientes->itens[i].id);
-    }
-    fclose(file);
-    printf("Paciente removido.\n");
-    
-}
-
-void update_pacientes(VetPacientes *pacientes){
-    
-    char choise;
-    int i, choise1;
-    Paciente paciente;
-    
-    i = search_paciente(pacientes);
-    
-    if(i == -1)return;
-    do{
-        printf("Deseja alterar aos dados desse paciente?(y/n)\n");
-        scanf(" %c", &choise);
-        if(choise == 'n')return;
-    }while(choise != 'y');
-    
-    do{printf("Digite o nome do paciente: ");
-        getchar();
-        fgets(paciente.nome,64, stdin);
-        paciente.nome[strcspn(paciente.nome, "\n")] = '\0';
-    }while(verify_name(paciente.nome)==0);
-    
-    do{printf("Digite o e-mail do paciente: ");
-    fgets(paciente.contato,64, stdin);
-    paciente.contato[strcspn(paciente.contato, "\n")] = '\0';
-    }while(verify_email(paciente.contato)==0);
-    
-    paciente.id = pacientes->itens[i].id;
-    
-    printf("Paciente atualizado\nNome: %s\nContato: %s\nId: %d\n", paciente.nome, paciente.contato,paciente.id);
-
-	FILE *file;
-    
-	file = fopen("pacientes.txt", "w");
-	if(file==NULL){
-		printf("Erro de memoria\n");
-		fclose(file);
-		return;
-	}
-	
-	pacientes->itens[i] = paciente;
-	
-	printf("Atualizando...\n\n");
-	for(i = 0; i < pacientes->qtd;i++){
-	    	fprintf(file, "%s | %s | %d \n",pacientes->itens[i].nome, pacientes->itens[i].contato, pacientes->itens[i].id );
-
-	}
-	fclose(file);
-}
-void list_pacientes(VetPacientes *pacientes){
-    for(int i = 0; i < pacientes->qtd; i++){
-        printf("id: %d nome: %s contato: %s\n", pacientes->itens[i].id, pacientes->itens[i].nome, pacientes->itens[i].contato);
-    }
-}
 
 int verify_name(const char *str){
     if (*str == '\0') return 0; // return 0 if the string is empty
@@ -231,28 +12,6 @@ int verify_name(const char *str){
     }
     return 1; //return 1 if the string is ok.
 }
-
-int verify_email(const char *str){
-    if(*str == '\0') return 0; // return 0 if str is empty
-    int qtd_arroba = 0, flag_arroba = -1, flag_ponto = -1, i; 
-    for(i = 0; str[i] != '\0'; i++){
-        if(str[i] == '@'){flag_arroba = i; qtd_arroba++;
-            if(i == 0){printf("e-mail inválido.\n"); return 0;} // the first char cannot be an "@"
-            else if(str[i-1] == '.' || str[i+1] == '.'){printf("e-mail inválido.\n"); return 0;} // the last char before "@" or the first one after "@" cannot be an "."
-        }
-        else if(str[i] == '.'){flag_ponto = i;
-            if(i == 0 || str[i+1] == '.'){printf("e-mail inválido.\n"); return 0;} // the first char cannot be an "." and a email cannot have two "." together.
-        }
-    }
-    
-    if(qtd_arroba != 1 || flag_ponto == -1 || flag_ponto < flag_arroba || str[i-1] == '.'){
-        printf("e-mail inválido.\n"); return 0; // e-mail only have 1 "@" and need at least 1 "." after the "@". The last char cannot be an "."
-    }
-    return 1;
-    
-}
-
-
 
 //Função para preencher horario mais facilmente
 void receber_hora(Horario *horario) {
@@ -365,597 +124,4 @@ int contar_linhas(FILE *arquivo) {
 	//Assim tem que usar o rewind para apontar o inicio do arquivo
 	rewind(arquivo);
 	return linhas;
-}
-
-//Para a todo momento verificar o tamanho do vetor, caso precise aumentar. Adiciona mais 10
-
-int realocar_medicos(VetMedicos *medicos) {
-	//Para comparar se a quantidade do vetor ocupado chegou no limite
-	if(medicos->qtd == medicos->cap) {
-
-		Medico *medicos1;
-		//declara um novo struct para garantir que o primeiro não vai se perder ao tentar realocar o espaço
-		medicos1 = (Medico*) realloc(medicos->itens, (medicos->cap+10) * sizeof(Medico));
-
-		if(medicos1 != NULL) {
-			//caso der certo, o vetor medicos.itens será realocado para ter o novo tamanho
-			medicos->itens = medicos1;
-			medicos->cap += 10;
-			return 1;
-		} else {
-			return 0;
-		}
-	}
-	return 1;
-}
-
-//Para converter o numero da especialidade em uma string
-//Especialidade é do tipo enum, ou seja, um inteiro
-const char* ler_especialidade(int n) {
-	//Recebe um numero da especialidade e devolve a string correta
-	switch(n) {
-	case 0:
-		return "Clinico";
-	case 1:
-		return "Pediatra";
-	case 2:
-		return "Dermatologista";
-	case 3:
-		return "Cardiologista";
-	case 4:
-		return "Outro";
-	default:
-		return "Especialidade invalida";
-	}
-
-}
-
-//Ler a quantidade de medicos no inicio da aplicação, para conseguir transcrever todos em um vetor
-//Converter todos medicos cadastrados no arquivo em um vetor
-void read_medicos(VetMedicos *medicos) {
-	FILE *file;
-
-	file = fopen("medicos.txt","r");
-
-	if(file == NULL) {
-		//Caso não exista arquivo nenhum, cria um vetor dinamico de tamanho 10
-
-		medicos->qtd = 0, medicos->cap = 10;
-		medicos->itens = (Medico *) malloc(sizeof(Medico) * medicos->cap);
-		//testa se há memaria para fazer a alocação
-		if(medicos->itens == NULL) {
-			printf("Erro de memoria\n");
-			return;
-		}
-
-		//cria um novo arquivo
-		file = fopen("medicos.txt","w");
-		if(file == NULL) {
-			printf("Erro ao abrir o arquivo\n");
-		}
-		fclose(file);
-		return;
-
-	} else {
-
-		//Chama uma função geral que serve para contar quantas linhas tem no arquivo,
-		//cria um vetor com a quantidade de linhas + 10.
-		//Considere cada linha um objeto
-
-		int n1, n2, n3, n4, i = 0;
-		//Variaveis n vão ser usadas para pegar os numeros inteiros que serão usado para pegar os hararios
-		medicos->qtd = contar_linhas(file);
-
-		medicos->cap = medicos->qtd + 10;
-		medicos->itens = (Medico *) malloc(sizeof(Medico) * medicos->cap);
-
-		//testa a alocação
-		if(medicos->itens == NULL) {
-			printf("Erro ao realocar o vetor\n");
-			fclose(file);
-			return;
-		}
-		//Varrer por todos elementos no arquivo, pegando cada atributo e colocando
-		//Usa | para dividir o elementos
-		//O %63[^|] garante pegar uma string com 63 elementos, ela vai lendo até parar no |
-		//Serve para pegar nomes compostos
-		while(fscanf(file, "%d|%63[^|]|%d|%d|%d|%d|%d",&medicos->itens[i].id, medicos->itens[i].nome, &medicos->itens[i].especialidade, &n1,&n2,&n3,&n4) ==7) {
-			//Adendo: strutc Especialidade é tipo enum, tem como representar como inteiro
-			//Como o struct Medico tem 4 variaveis Horario
-			//Chamo as funções desconverter_horas para já prencher cada espaço da variavel. Otimizando o processo
-			desconverter_horas(n1,&medicos->itens[i].inicioManha);
-			desconverter_horas(n2,&medicos->itens[i].fimManha);
-			desconverter_horas(n3,&medicos->itens[i].inicioTarde);
-			desconverter_horas(n4,&medicos->itens[i].fimTarde);
-			i++;
-		}
-		//Por garantia pega agora o numero de fato que existe em medicos
-		medicos->qtd = i;
-		fclose(file);
-		return;
-	}
-}
-//Para quando necessario mostrar os medicos
-void mostrar_medico(Medico medico) {
-	printf("Medico: %s      id: %d      Especildiade: %s\n", medico.nome,medico.id, ler_especialidade(medico.especialidade));
-	printf("Horario de manhâ das %d:%d até %d:%d\n",medico.inicioManha.horas, medico.inicioManha.minutos, medico.fimManha.horas, medico.fimManha.minutos);
-	printf("Horario de tarde das %d:%d até %d:%d\n",medico.inicioTarde.horas, medico.inicioTarde.minutos, medico.fimTarde.horas, medico.fimTarde.minutos);
-
-}
-void add_medico(VetMedicos *medicos) {
-	int maior = 0, choise1;
-	char choise;
-	//medico é usado como uma variavel de espaço temporario, só quando é confirmado adição que os dados são guardados
-	Medico medico;
-
-	//testa se é necessario aumentar o vetor de medicos
-	choise1 = realocar_medicos(medicos);
-
-	if(choise1==0) {
-		printf("Não há mais memória para armazenar novos medicos. Encerrando função adicionar.\n");
-		return;
-	}
-
-	//Percorre todo vetor de medicos, no final pega o maior id existente e adiciana +1. Caso não haja nenhum medico, o novo recebe 1
-	if(medicos->qtd > 0) {
-		for(int i=0; i<medicos->qtd; i++) {
-			if(medicos->itens[i].id>=maior) {
-				maior = medicos->itens[i].id;
-			}
-		}
-	}
-	medico.id = maior+1;
-    do{
-        printf("Digite o nome do medico: ");
-    	getchar();
-    	fgets(medico.nome,64, stdin);
-    	medico.nome[strcspn(medico.nome, "\n")] = '\0';
-    }while(verify_name(medico.nome)!= 1);
-	
-	//Recebe a especialidade do medico e converte para ficar da maneira correta
-	while(1) {
-		printf("Escolha uma das seguintes especialidade do médico\nClinico: 1\nPediatra: 2\nDermatologista: 3\nCardiologista: 4\nOutra: 5\n");
-		scanf("%d",&choise1);
-		if(choise1>=1 && 5>=choise1) {
-			break;
-		}
-		else {
-			printf("Digite uma opção válida\n");
-		}
-	}
-
-	medico.especialidade = choise1 - 1;
-
-	printf("Adendo: cada turno do medico vai ser calculado com 3 horas de duração\n");
-
-	printf("    Preencha o horário do expediente do inicio da manhã\n");
-	receber_hora(&medico.inicioManha);
-	veri_horario(&medico.inicioManha,&medico.fimManha,1, 1);
-	
-	printf("    Preencha o horário do expediente do inicio da tarde\n");
-	receber_hora(&medico.inicioTarde);
-	veri_horario(&medico.inicioTarde,&medico.fimTarde,1, 2);
-
-	printf("        Deseja adicionar\n");
-	mostrar_medico(medico);
-
-	while(1) {
-		printf("Digite y para confirmar ou n para cancelar\n");
-		scanf(" %c", &choise);
-		if(choise == 'n')return;
-		if(choise == 'y')break;
-	}
-	//Cria agora o arquivo, caso fosse no inicio e o usuario não criasse o medico, o arquivo não seria fechado. Assim podendo corromper o arquivo
-	FILE *file;
-
-	file = fopen("medicos.txt", "a");
-	if(file==NULL) {
-		printf("Erro de memoria\n");
-		return;
-	}
-	printf("Adicionando...\n\n");
-	fprintf(file, "%d|%s|%d|%d|%d|%d|%d\n",medico.id, medico.nome, medico.especialidade,con_horas(medico.inicioManha),con_horas(medico.fimManha),con_horas(medico.inicioTarde),con_horas(medico.fimTarde));
-
-	fclose(file);
-
-	//Como medicos->itens é do tipo Medico, ele consegue receber diretamente a variavel medico
-	medicos->itens[medicos->qtd] = medico;
-
-	medicos->qtd++;
-}
-
-int pesquisar_medicos(VetMedicos *medicos) {
-	int numero, i;
-	Medico medico1;
-	
-	if(medicos->qtd < 1){printf("Não existe medicos cadastrados.\n");return -2;}
-	
-	printf("Digite o id do medico desejado\n");
-	scanf("%d", &numero);
-	//Percorre todo vetor de medicos até encontrar o medico com id escolhido pelo usuario.
-
-	for(i = 0; i < medicos->qtd && numero!=medicos->itens[i].id; i++);
-	//Testa se encontrou o medico
-	if( i < medicos->itens[i].id && numero!=0) {
-		if(numero == medicos->itens[i].id) {
-			printf("Medico encontrado\n");
-			mostrar_medico(medicos->itens[i]);
-			return i;
-		}
-	} else {
-		printf("Medico não encontrado no sistema\n");
-		return -1;
-	}
-}
-
-void update_medicos(VetMedicos *medicos){
-    char choise;
-    int i, choise1;
-    Medico medico;
-    
-    i = pesquisar_medicos(medicos);
-    
-    if(i == -1)return;
-    do{
-        printf("Deseja alterar aos dados desse medico?(y/n)\n");
-        scanf(" %c", &choise);
-        if(choise == 'n')return;
-    }while(choise != 'y');
-    
-    do{
-        printf("Digite o nome do medico: ");
-    	getchar();
-    	fgets(medico.nome,64, stdin);
-    	medico.nome[strcspn(medico.nome, "\n")] = '\0';
-    }while(verify_name(medico.nome)!= 1);
-	
-	while(1) {
-		printf("Escolha uma das seguintes especialidade do médico\nClinico: 1\nPediatra: 2\nDermatologista: 3\nCardiologista: 4\nOutra: 5\n");
-		scanf("%d",&choise1);
-		if(choise1>=1 && 5>=choise1) {
-			break;
-		}
-		else {
-			printf("Digite uma opção válida\n");
-		}
-	}
-
-	medico.especialidade = choise1 - 1;
-
-	printf("Adendo: cada turno do medico deve ser de exatamente 3 horas de duração\n");
-
-	printf("    Preencha o horário da consulta do inicio da manhã\n");
-	receber_hora(&medico.inicioManha);
-	veri_horario(&medico.inicioManha,&medico.fimManha,1, 1);
-	printf("    Preencha o horário de consulta do inicio da tarde\n");
-	receber_hora(&medico.inicioTarde);
-	veri_horario(&medico.inicioTarde,&medico.fimTarde,1, 2);
-
-	printf("Medico atualizado\n");
-	mostrar_medico(medico);
-
-	FILE *file;
-    
-	file = fopen("medicos.txt", "w");
-	if(file==NULL) {
-		printf("Erro de memoria\n");
-		fclose(file);
-		return;
-	}
-	
-	medico.id = medicos->itens[i].id;
-	medicos->itens[i] = medico;
-	
-	printf("Atualizando...\n\n");
-	for(i = 0; i < medicos->qtd;i++){
-	    	fprintf(file, "%d|%s|%d|%d|%d|%d|%d\n",medicos->itens[i].id, medicos->itens[i].nome, medicos->itens[i].especialidade,con_horas(medicos->itens[i].inicioManha),con_horas(medicos->itens[i].fimManha),con_horas(medicos->itens[i].inicioTarde),con_horas(medicos->itens[i].fimTarde));
-
-	}
-	fclose(file);
-    
-}
-void remover_medico(VetMedicos *medicos) {
-	int i, id;
-	char choise;
-	id = pesquisar_medicos(medicos);
-	//Aproveita a função anterior para já pegar o id do medico selecionado para ser deletado
-	if(id==-1)return;
-	printf("Deseja remover esse medico?\n");
-
-	while(1) {
-		printf("Digite y para confirmar ou n para cancelar\n");
-		scanf(" %c", &choise);
-		if(choise == 'n')return;
-		if(choise == 'y')break;
-	}
-
-	FILE *file;
-	//Remove o medico
-	//Basicamente joga todos elemetos a direita dele para esquerda. Sobreescrevendo o elemento apagado
-	for(i = id; i < medicos->qtd-1 ; i++) {
-		medicos->itens[i] = medicos->itens[i + 1];
-	}
-	medicos->qtd--;
-
-	file = fopen("medicos.txt", "w");
-	//Limpa todo o arquivo e em seguida preenche novamente com os dados restantes do vetor
-	for(i = 0; i<medicos->qtd; i++) {
-		fprintf(file, "%d|%s|%d|%d|%d|%d|%d\n",medicos->itens[i].id, medicos->itens[i].nome, medicos->itens[i].especialidade,con_horas(medicos->itens[i].inicioManha),con_horas(medicos->itens[i].fimManha),con_horas(medicos->itens[i].inicioTarde),con_horas(medicos->itens[i].fimTarde));
-	}
-
-	fclose(file);
-	printf("Removendo medico...\n");
-}
-//Lista medicos filtrando por especialidade e mostra a quantidade de medicos cadastrados
-void listar_medicos(VetMedicos *medicos) {
-	int choise,i, flag = 0;
-	char choise1;
-
-	printf("Quantidade total de medicos cadastrados no sistema: %d\n", medicos->qtd);
-
-    while(1) {
-			printf("1 - listar todos\n2 - especialidade especifica\n");
-			scanf("%d",&choise);
-			if(choise==1 || 2==choise) {
-				break;
-			}
-			else {
-				printf("Digite uma opção válida\n");
-			}
-		}
-	if(choise == 1){
-	    for(i = 0; i<medicos->qtd; i++) {
-			mostrar_medico(medicos->itens[i]);
-			printf("\n");
-		}
-	}else{
-	    while(1) {
-    		//Pede a especialidade do medico e depois converte para a função ler_especialidade retornar um valor valido
-    		while(1) {
-    			printf("Escolha uma das seguintes especialidade do médico para filtrar e listar\nClinico: 1\nPediatra: 2\nDermatologista: 3\nCardiologista: 4\nOutra: 5\n");
-    			scanf("%d",&choise);
-    			if(choise>=1 && 5>=choise) {
-    				break;
-    			}
-    			else {
-    				printf("Digite uma opção válida\n");
-    			}
-    		}
-    
-    		choise--;
-    
-    		for(i = 0; i<medicos->qtd; i++) {
-    			if(choise == medicos->itens[i].especialidade) {
-    				mostrar_medico(medicos->itens[i]);
-    				printf("\n");
-    				flag++;
-    			}
-    		}
-    		//Caso não exista um medico com essa especialidade, retorna uma mensagem
-    		if(flag==0) {
-    			printf("Nenhum medico dessa especialidade cadastrado\n");
-    			while(1) {
-    				printf("Deseja filtrar por outra especialdiade?\nDigite y para confirmar ou n para cancelar\n");
-    				scanf(" %c", &choise);
-    				if(choise == 'n')return;
-    				if(choise == 'y')break;
-    			}
-    		}
-    		else {
-    			return;
-    		}
-    	}
-	}
-	
-}
-
-void add_consulta(VetConsultas *consultas, VetPacientes *pacientes, VetMedicos *medicos){
-    int status = 0; //teste
-    int id_paciente, id_medico;
-    char choice;
-    
-    
-    do{id_paciente = search_paciente(pacientes);if(id_paciente==-2)return;}while(id_paciente==-1);
-    do{id_medico = pesquisar_medicos(medicos);if(id_medico==-2)return;}while(id_medico==-1);
-    Data data; Horario inicio, fim;
-    add_data(&data);
-    printf("Horario de inicio. ");
-    receber_hora(&inicio); 
-    veri_horario(&inicio, &fim, 2, 0);
-    
-    printf("agendar consulta? (y/n)");
-    while(1){
-        scanf(" %c", &choice);
-    
-    if(choice == 'n') return;
-    if(choice=='y')break;
-    }
-    
-    int maior = 0;
-    if(consultas->qtd > 0){
-        for(int i = 0; i < consultas->qtd; i++){
-            if(consultas->itens[i].id > maior) maior = consultas->itens[i].id;
-        }
-    }
-    
-    FILE *file;
-    
-    file = fopen("consultas.txt", "a");
-    
-    fprintf(file, "%d | %d | %d | %d | %d | %d | %d\n", maior+1, pacientes->itens[id_paciente].id, medicos->itens[id_medico].id, con_data(data), con_horas(inicio), con_horas(fim), status);
-    
-    fclose(file);
-    
-    //feeding the vector:
-    
-    consultas->itens[consultas->qtd].id = maior+1;
-    consultas->itens[consultas->qtd].idPaciente = pacientes->itens[id_paciente].id;
-    consultas->itens[consultas->qtd].idMedico = medicos->itens[id_medico].id;
-    consultas->itens[consultas->qtd].data.dia = data.dia;
-    consultas->itens[consultas->qtd].data.mes = data.mes;
-    consultas->itens[consultas->qtd].data.ano = data.ano;
-    consultas->itens[consultas->qtd].inicio.horas = inicio.horas;
-    consultas->itens[consultas->qtd].inicio.minutos = inicio.minutos;
-    consultas->itens[consultas->qtd].fim.horas = fim.horas;
-    consultas->itens[consultas->qtd].fim.minutos = fim.minutos;
-    consultas->itens[consultas->qtd].status = 0;
-    
-    consultas->qtd++;
-    printf("Consulta agendada.\n");
-    
-}
-
-const char* ler_status(int n) {
-	//Recebe um numero do status e devolve a string correta
-	switch(n) {
-	case 0:
-		return "CONS_AGENDADA";
-	case 1:
-		return "CONS_CONCLUIDA";
-	case 2:
-		return "CONS_CANCELADA";
-	case 3:
-		return "CONS_FALTA";
-	default:
-		return "Status invalido";
-	}
-
-}
-void printar_consultas(Consulta *consultas){
-    printf("id: %d id_medico: %d id_paciente: %d data: %d/%d/%d horario: %dh%d as %dh%d status: %s\n", consultas->id, consultas->idMedico, consultas->idPaciente, consultas->data.dia, consultas->data.mes, consultas->data.ano, consultas->inicio.horas, consultas->inicio.minutos, consultas->fim.horas, consultas->fim.minutos, ler_status(consultas->status));
-}
-
-void list_consultas(VetConsultas *consultas){
-    int choice = 0;
-    printf("1 - Todas\n2 - Agendadas\n3 - Concluidas\n4 - Canceladas\n5 - Faltas\n");
-    do{scanf("%d", &choice);}while(choice > 5 || choice < 1);
-    if(choice == 1){
-        for(int i = 0; i<consultas->qtd; i++){
-            printf("id: %d id_medico: %d id_paciente: %d data: %d/%d/%d horario: %dh%d as %dh%d status: %s\n", consultas->itens[i].id, consultas->itens[i].idMedico, consultas->itens[i].idPaciente, consultas->itens[i].data.dia, consultas->itens[i].data.mes, consultas->itens[i].data.ano, consultas->itens[i].inicio.horas, consultas->itens[i].inicio.minutos, consultas->itens[i].fim.horas, consultas->itens[i].fim.minutos, ler_status(consultas->itens[i].status));
-        }
-    }else{
-        for(int i = 0; i<consultas->qtd; i++){
-            if(consultas->itens[i].status == (choice-2))printf("id: %d id_medico: %d id_paciente: %d data: %d/%d/%d horario: %dh%d as %dh%d status: %s\n", consultas->itens[i].id, consultas->itens[i].idMedico, consultas->itens[i].idPaciente, consultas->itens[i].data.dia, consultas->itens[i].data.mes, consultas->itens[i].data.ano, consultas->itens[i].inicio.horas, consultas->itens[i].inicio.minutos, consultas->itens[i].fim.horas, consultas->itens[i].fim.minutos, ler_status(consultas->itens[i].status));
-    }}
-}
-
-int con_data(Data data){
-    data.ano *= 10000;
-    data.mes *= 100;
-    return data.dia + data.mes + data.ano;
-}
-void add_data(Data *data){
-    while(1){
-        printf("Digite um dia para consulta:\n");
-        scanf("%d", &data->dia);
-        if(30 >= data->dia && data->dia >=1){
-            break;
-        }else{printf("Erro, digite um dia valido\n");}
-    }
-    while(1){
-        printf("Digite um mes para consulta:\n");
-        scanf("%d", &data->mes);
-        if(12 >= data->mes && data->mes >=1){
-            break;
-        }else{printf("Erro, digite um mes valido\n");}
-    }
-    while(1){
-        printf("Digite um ano para consulta:\n");
-        scanf("%d", &data->ano);
-        if(data->ano >=2026){
-            break;
-        }else{printf("Erro, digite um ano valido\n");}
-    }
-}
-void descon_data(Data *data, int n){
-    data->ano = n / 10000; 
-    data->mes = n % 10000 / 100;
-    data->dia = n % 10000 % 100;
-}
-
-void read_consultas(VetConsultas *consultas){
-	FILE *file;
-
-	file = fopen("consultas.txt","r");
-
-	if(file == NULL) {
-
-		consultas->qtd = 0, consultas->cap = 10;
-		consultas->itens = (Consulta *) malloc(sizeof(Consulta) * consultas->cap);
-		if(consultas->itens == NULL) {
-			printf("Erro de memoria\n");
-			return;
-		}
-
-		//cria um novo arquivo
-		file = fopen("consultas.txt","w");
-		if(file == NULL) {
-			printf("Erro ao abrir o arquivo\n");
-		}
-		fclose(file);
-		return;
-
-	} else {
-
-		int n1, n2,n3, i = 0;
-
-		consultas->qtd = contar_linhas(file);
-
-		consultas->cap = consultas->qtd + 10;
-		consultas->itens = (Consulta *) malloc(sizeof(Consulta) * consultas->cap);
-
-		if(consultas->itens == NULL) {
-			printf("Erro ao realocar o vetor\n");
-			fclose(file);
-			return;
-		}
-		
-		while(fscanf(file, "%d | %d | %d | %d | %d | %d | %d",&consultas->itens[i].id, &consultas->itens[i].idMedico, &consultas->itens[i].idPaciente, &n1,&n2,&n3, &consultas->itens[i].status) ==7) {
-			descon_data(&consultas->itens[i].data, n1);
-			desconverter_horas(n2,&consultas->itens[i].inicio);
-			desconverter_horas(n3,&consultas->itens[i].fim);
-			i++;
-		}
-		
-		consultas->qtd = i;
-		fclose(file);
-		return;
-	}
-}
-
-int realocar_consultas(VetConsultas *consultas) {
-	if(consultas->qtd == consultas->cap) {
-		Consulta *consultas1;
-		consultas1 = (Consulta*) realloc(consultas->itens, (consultas->cap+10) * sizeof(Consulta));
-		if(consultas1 != NULL) {
-			consultas->itens = consultas1;
-			consultas->cap += 10;
-			return 1;
-		} else {
-			return 0;
-		}
-	}
-	return 1;
-}
-
-int search_consultas(VetConsultas *consultas){
-    int numero, i;
-	printf("Digite o id da consulta desejado\n");
-	scanf("%d", &numero);
-    
-    printf("Quantidade de consultas: %d\n", consultas->qtd);
-	for(i = 0; i < consultas->qtd && numero!=consultas->itens[i].id; i++);
-
-	printf("ID de consultas: %d\n", consultas->itens[i].id);
-
-	if( i < consultas->itens[i].id && numero!=0) {
-		if(numero == consultas->itens[i].id) {
-			printf("Consulta encontrado\n");
-			printar_consultas(&consultas->itens[i]);
-			return i;
-		}
-	} else {
-		printf("Consulta não encontrado no sistema\n");
-		return -1;
-	}
 }
